+++ FrozenYoghourt/maths/__init__.py
@@ -1,6 +1,7 @@
 from FrozenYoghourt import *
 from FrozenYoghourt.mode import *
 
+close = lambda a, b: np.all(np.isclose(a, b))
 
 def mm(*lst, no_times=1):
     lst *= no_times